lots of changes, still getting code ready for the timer for level 2 dispensers

+++ src/main/java/net/goldcoops/foliatntgenfix/commands/GiveDuplicatorCommand.java
@@ -25,7 +25,7 @@ public class GiveDuplicatorCommand implements CommandExecutor {
 
         if (args.length > 0) {
 
-            ItemStack tntduper = createTntDispenserItem(1);
+            ItemStack tntduper = createTntDispenserItem((byte) 1);
 
             String targetName = args[0];
